requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
langgraph>=0.0.40
sqlalchemy>=2.0.0